requirements.txt
streamlit
scilit-learn
pickles
streamlit-lottie
pytesseract
open-cv
pillow
numpy
deep-translator 